[CMake] Add tests and examples to ctest
Add all tests currently called from the run_* scripts in the tests
directory to CMake, so they can be built automatically and run using
'ctest' in the build directory. Set ENABLE_TESTS=ON to enable this.

This also changes some paths so that all build artefacts end up inside the build
directory, and no longer pollute the source tree.