CMake: Add a package config
Instead of consuming projects needing to define a find module for occa,
a package config is provided by occa itself. Thus all occa targets are
correctly imported during a find_package(occa), including the
information on their dependencies. And in the package config it is
defined what targets are being made available.

Consumers can then go directly to
target_link_libraries(theirTarget OCCA::libocca).

The occa binary is also exported as a target, because a consuming
project may want to use it in e.g. a custom command.