Cmake dependency checking: consistent use of find_package
Why: In order to prepare for adding a package config for occa itself,
custom dependency finding code should be in (a) file(s) that can be
installed alongside the package config file.
Also, to have a consistent way in which we find dependencies.

Therefore: add Find modules for all dependencies for which CMake does
not provide one, or where we have additional finding code that needs to
be wrapped. And, always use find_package within the main CMakeLists to
find a dependency.