Package config with Find module wrappers (#533)
* CmakeLists.txt: add include dir to install interface

When building against libocca, projects will need the include files.

* Cmake dependency checking: consistent use of find_package

Why: In order to prepare for adding a package config for occa itself,
custom dependency finding code should be in (a) file(s) that can be
installed alongside the package config file.
Also, to have a consistent way in which we find dependencies.

Therefore: add Find modules for all dependencies for which CMake does
not provide one, or where we have additional finding code that needs to
be wrapped. And, always use find_package within the main CMakeLists to
find a dependency.

* CMake: Add a package config

Instead of consuming projects needing to define a find module for occa,
a package config is provided by occa itself. Thus all occa targets are
correctly imported during a find_package(occa), including the
information on their dependencies. And in the package config it is
defined what targets are being made available.

Consumers can then go directly to
target_link_libraries(theirTarget OCCA::libocca).

The occa binary is also exported as a target, because a consuming
project may want to use it in e.g. a custom command.

* Cmake: Improve relocatability by using imported targets

For Threads, OpenMP, (CUDA,) OpenCL and MPI, use the IMPORTED targets provided by the
Cmake Find modules, instead of path-specific variables, to link to.
For DPCPP, HIP and METAL, create an IMPORTED INTERFACE library ourselves, as it is not provided.

* Cmake: fix casing of AND in if

* Cmake: Only look for OPENCL_ROOT if not already set

* Cmake: FindHIP fix compile definition

The target_property INTERFACE_COMPILE_DEFINITIONS should not have a leading -D, only the define itself.

Co-authored-by: Noel Chalmers <noel.chalmers@gmail.com>

Co-authored-by: Noel Chalmers <noel.chalmers@gmail.com>