Fix segfault when using --verbose flag when running examples with Intel mpiexec
Since updating to the Intel 2021.2 compiler stack, mpi-related tests started
failing with a segmentation fault in MPI_Init.

The '--verbose' flag passed to our example executables seems to have a bad
interaction with mpiexec: if we also pass '--verbose' (or '-verbose') to
mpiexec the example runs wtihout a crash.

If we use mpirun instead of mpiexec the problem also goes away, but CMake
sets MPIEXEC_EXECUTABLE to mpiexec, so it's better to keep it that way.

Luckily, none of the mpi-related examples actually parse '--verbose', so
we can actually leave it out without any detrimental effects.