cmake: Small fix to prevent empty module path
In a chain of cmake package config projects using find_dependency(),
where each would use the same code to temporarily adapt the module
path, the path would end up empty in the end, even if it wasn't empty
before, because of the unset() call of a not-package-specific variable.
Therefore, make the variable name package specific.