Add env var to override OCCA HIP compile arch
This commit teaches OCCA to understand the OCCA_HIP_OVERRIDE_COMPILE_ARCH
environment variable.  If set, OCCA will ignore the arch set in the kernel
properties and try to compile for the architecture provided in the environment
variable.

This allows the user to have more control over what targets are JITted for.
Usecases are for JITting for generic targets (see here:
https://github.com/llvm/llvm-project/commit/f93aa5157a3317b24cff660ac972814ee9ed4dbc)
and for the (currently beta) amdgcnspirv target (see here:
https://rocm.docs.amd.com/projects/llvm-project/en/latest/conceptual/spirv.html#rocm-support-for-spir-v-beta).

Unfortunately, the amdgcnspirv target name doesn't start with "gfx", so I added
a branch that checks for this case and treats it accordingly.